+++ app/models.py
@@ -120,3 +120,13 @@ def stripfilename(filename):
 	noextension=os.path.splitext(basename)[0] 
 	separateCamelCase = re.sub(r'((?<=[a-z])[A-Z]|(?<!\A)[A-Z](?=[a-z]))', r' \1', noextension)
 	return separateCamelCase.replace("_"," ")
+
+def get_or_create(model, **kwargs):
+    instance = db.session.query(model).filter_by(**kwargs).first()
+    if instance:
+        return instance
+    else:
+        instance = model(**kwargs)
+        db.session.add(instance)
+        db.session.commit()
+        return instance
