+++ tmp/soundboard.log
@@ -1,0 +1,1 @@
+2013-11-22 16:49:19,651 INFO: soundboard startup [in H:\Documents\Github\simplesoundboard\app\__init__.py:30]
